Updated openstack/openstack Project: openstack-infra/storyboard-webclient  55712fc403160d86fe6a2bf7057d936022689a17
Display a modal when a card is clicked on

This modal shows some information about the task/story the
card represents, and also allows it all to be edited. This
includes changing task assignee and assigning due dates to
the task.

Only one due date can be displayed on a card at any given
time.

Change-Id: I69866756e4b6b569e7f88cf41bb00c45ca5dedb9