Updated openstack/openstack Project: openstack/nova  4ebf5b469a8ccf527a9e88f9754278e1aeca6830
Allocate uuids for aggregates as they are created or loaded

Note that this filters uuid from the aggregate view in the API because we would
need a microversion. That may be a thing in the future (for the 2.x api) but
not something we can do here.

Related to blueprint generic-resource-pools

Change-Id: I45006e546867d348563831986b91a317029a1173