Updated openstack/openstack Project: openstack/barbican  3006c7edc497b577eb4ea7856ee0b9331c4a1691
Casts algorithm to lower case in KMIP plugin.

The KMIP plugin verifies if an algorithm that was given by
the end user is supported by the KMIP plugin.After verification,
the algorithm needs to be casted to lower case.If not, then
an error will occur and the KMIP plugin will fail to generate
the secret.

Change-Id: I0b21b0310e5aa445045258d3a89f2dbedf2809ef
Closes-Bug: 1537182