Updated openstack/openstack Project: openstack/nova  332db7db350ff5f77e9d5c6177e5e4512eed0a0b
add a place for functional test to block specific regressions

We get some bugs reported in by users which we can replicate and fix
most easily in a full stack way. Having a dedicated place to put these
seems like a good idea. They are in no way an attempt to test all the
things, they are specific tests that demonstrate a bug, and ensure
that once that particular bug is fixed, it will not come back.

Change-Id: Idfd8680133197beabaf6e1ac8df7490c2a345b17
Related-Bug: #1522536