Updated openstack/openstack Project: openstack-infra/project-config  132c8b03562f3d891ba049b7e8e84d4dfe31cd49
Also purge the hiera package that comes preinstalled in dsvm

puppet-prepare-node already took care of uninstalling the
puppet and facter packages that came preinstalled with the dsvm.
Hiera was not uninstalled and was shown to cause some issues when
attempting to use it.

Let's also uninstall hiera, it will be re-installed inside the jobs
when installing the puppet package, it does not need to be
specified explicitely as it comes bundled with puppet now.

Change-Id: I5e03d3de3c28afffb8dd30e429197eef02385520