Updated openstack/openstack Project: openstack/nova  01f5c73bf51e9aec69c23a730d4021b91d06fd8c
make microversion a client level construct for tests

Instead of passing around microversion in every method call, the way
microversions are expected to be used by a client is by setting it
globally for the client and using that microversion for all requests.

This adds that support to the base test client, and updates nearly all
tests to use it that way. test_server_group is tightly coupled and
will require some much more specific unwind.

Change-Id: I2263104c34328f11c687c6b8ef3a1cfd21980a63