Updated openstack/openstack Project: openstack-dev/devstack  6c888d2c49613f94c3325f2474bfa71a1de3061a
localize user/project variables in tempest

lib/tempest is using / setting a ton of globals, a lot which don't
look used. As a first step to converting over to s/tenant/project/
make all of these local which means they can not be impacting anything
outside of this function. That will make them safe to change names
later.

Change-Id: I380dd20e5ed716a0bdf92aa02c3730359b8136e4