Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  d30516dfe28aaa83ec08251f00894371a246575b
Allow null values for log-text

The log-text parameter is required according to the documentation,
but it's perfectly valid for it to be an empty string or null.

If the value is null in the yaml, the Python data structure
contains None. The code was converting this directly to a string,
resulting in 'None' in the xml output.

This change makes sure that an empty string is used instead,
resulting in a correct empty element in the xml.

Change-Id: Ided872f5528bae93347f96bd729b5ce37f8e16fb