Updated openstack/openstack Project: openstack/cinder  68e7ae469d9dc1f907e01ba39e441b61475a232a
Fix HNAS iSCSI driver attachment

Currently, the HNAS iSCSI driver returns the logical unit number
related to the iSCSI target as a string while initializing the
connection.
This is breaking the attachment feature because os-brick expects
a numeric type.
This patch converts the LUN from string to int in order to solve
this problem.

Change-Id: I33958a93a6261b7d77e4becafa3efd5180e5d091
Closes-Bug: 1549375