Updated openstack/openstack Project: openstack/glance  5be9a71ec4207b13f72e9b3c3db78adc27958285
clean up auto-generated docs for configuration options

Split each config file out to its own page to avoid the SEVERE errors
from duplicate targets on a given page.

Use the new ability of oslo.config.sphinxext to read the existing
configuration files for the config generator to load the set of
namespaces for a given file to show all options for the service.

Change-Id: I844f43049b2b6def01541a8d761cebd0c2b0b324
Depends-On: I0ec46cc7aa820ae592d8e4e7b56ee98f0c5cbfb4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>