Updated openstack/openstack Project: openstack-infra/system-config  c57170d23823d904559259da97a5c8aed2ca8c4d
Playbook for setting the infracloud jenkins project quotas

We are only deploying West for now, so just doing West.
When we get East in production, we would update this playbook.
Unfortunate there is no Ansible module or Puppet resources to set
quotas per-project, thus using regular shell module in Ansible.

Change-Id: Ib884508bebedc9f88fac242711af98fc0c4d95ec