Updated openstack/openstack Project: openstack-infra/system-config  c837752f3f57a9aeaa4f9f52df12152d1b51fd8b
Instruct puppet to configure settings before applying

In a clean deploy of infra cloud, the puppet environment
is not configured from scratch. That will prevent puppet to run
because it won't find the /opt/system-config/production/modules.

The config option of the ansible role will configure properly
all settings before trying to apply it, and things will work
properly.

Change-Id: I736e10623fb3ba90b3320cc20758a18c70930be0
Depends-On: I6cb8dff569f2cca8bca7359412d01cc7ec009c54