Updated openstack/openstack Project: openstack/nova  8974a30914ae26c0b1bb9386e46af877de737f61
Make online_data_migrations do smaller batches in unlimited case

We want to have a single command that people can run to do all of their
online migrations in a single go. This is the closest thing to the
old model of "run all my database migrations, let me know when it's
done." The code right now ends up doing a single batch of unlimited size,
which will undoubtedly run out of memory for large enough deployments.

This patch makes us execute smaller batches until we're done. If
people want to optimize this, they can control the batch size and
do their own looping, but we need a reasonable option for people
that just want one-shot without the potential for the OOM.

Change-Id: I07ec2aeb0c305dd1a088ca590f3c91a8ced10787