Updated openstack/openstack Project: openstack/nova  29dc5df839d3019716396aee2d97599426723337
Update the CONF import path for VNC

The VNC conf options were recently moved to nova.conf.vnc as part of the
centralize-config-options blueprint.  This changes the import path for
the oslo import_opt command.

Most of the code just work because nova pre-imported the nova.conf
package.  But we found in the out-of-tree powervm driver that the
import_opt command needed to change the group to nova.conf.vnc to work.

The appropriate fix is instead of importing cfg.CONF, to have the code
import nova.conf.CONF.  This properly resolves all of the imports.

Change-Id: I36d767be527e655a89b7635f92433003a888ccbc
Implements: bp centralize-config-options