Updated openstack/openstack Project: openstack/barbican  83e7caa02c571df1d7a4c86f45493eb7afaf0c4e
Fix roles attribute for barbican request context

The oslo_context.context.RequestContext was updated in oslo_context 2.2.0
to have a roles attribute. This caused the barbican request context to
have an empty roles attribute.
The 'inspect' library is used to make sure backwards compatibility since
the current requirements for oslo_context is >= 0.2.0.

Change-Id: Iafc67a6f42e46da998eaf2f5199c47f6beaf345e