Updated openstack/openstack Project: openstack/python-heatclient  fdc1ee7deeff894fc37ceb5cf6791910af3ed3af
Changes to support server-side environment resolution

Adds support in template_utils to optionally produce an ordered list of
environment file names and include the environment file contents in the
files dict. This will be passed to the server to trigger the server-side
handling of resolving overrides for multiple environments.

The original behavior of merging the environments client-side is left in
for backward compatibility purposes. Both are sent to the server. For
mitaka and newer, the client-submitted environment is ignored and the
merging is done server-side. For liberty and older, the
environment_files list is ignored entirely and the passed in environment
is used.

Note that the merging, regardless if it is done client or server-side,
is handled by the same code; it has not been copied or duplicated in the
server.

Change-Id: If187c8ca3d80008c21e8f6066d335ab0c30315e6
Implements: blueprint multi-environments
Depends-On: Ibe46fd35de7988920c101a83259c06c8f8a3ed0b