Updated openstack/openstack Project: openstack/horizon  11795a8c5c46994ffe241b8ccd0a8a7b493f79e0
Adding features to MagicSearch

The previous MagicSearch patch was a function-for-function copy of
the source at GitHub.  There are some useful features that were missing
from that.

This patch adds a 'clientFullTextSearch' property to the magic search bar so
the full-text search option may be turned off.

Change-Id: Id46b169c5ba18606d498701b6664fc6374083011
Implements: blueprint integrate-magic-search