Updated openstack/openstack Project: openstack-dev/devstack  447af8f3f60e407679ff3a13d16db7c5183e12f6
Generate plugins list through https requests

The devstack plugins list can be generated through web requests in
environments (such as the proposal slave) that lack copies of all
the relevant git repositories.

One downside to this is that there is no way of getting the last
modification time of the plugin.

Change-Id: I2c5c9282a8ad80014cad171a4dfbdc8f26044cd1