Updated openstack/openstack Project: openstack-infra/project-config  af308f7129e2e2314ae393e1fc45eb1e3b45225b
Revert "Add allow-local-ssh-root to the packstack integration jobs"

This reverts commit 7301e565ae567aad8f4aa72cab2164c7750d5f51.

There are unfortunately several hacks involved to make Packstack work
properly in the gate. We thought this would work but it does not.
We will leave the SSH connection setup logic inside the run_tests
script which was proven to work successfully.

Change-Id: I41499519f2c81990ba1112eb1ded959970ab998a