Updated openstack/openstack Project: openstack/sahara-dashboard  6ba85222015cb43fd591e72f1a1556ecd7e63a61
Changing name of "Data image registry"

The name "Data image registry" is not very good.
Given that the image registry appears under the
data processing panels, it should be obvious that
it applies to data processing.  Therefore, I am removing
the "data" part of the name.  Also updating the
integration test that references this page and
cleaning up unused code from views.py.

Change-Id: Ie60c8e3154ae9f640f154ca26d42c0c0a569b9a5
Partial-Implements: bp reduce-number-of-panels