Updated openstack/openstack Project: openstack/nova  36a4fe61a782695c37ad155dd883b048853b706f
fix usage of opportunistic test cases with enginefacade

This creates an in tree fixture that lets you rebuild the engine
facade for doing individual tests against MySQL. It is needed until
this is supported in upstream oslo.db, which might be some time given
we are just hitting library freeze. Without this we can't build
regression tests for mysql specific bugs.

Related-Bug: #1548960

Change-Id: I2c4148972e91dcf1d6df0960198dc842b633d9e7