Updated openstack/openstack Project: openstack-infra/tripleo-ci  2c16607140f37e7396db0be16e27d63455210695
Redirect undercloud install output to file

The undercloud install is not quite as verbose as the image build,
but it's in the same ballpark and it makes it take a long time to
load the console log from our test runs.  Since most of our failures
happen as part of overcloud deployment, let's send the undercloud
install output to a log file so it's still available, but doesn't
have to be downloaded every time you want to look at the output
from a test run.

Change-Id: I2a50d8d7f0034a19de368bdad66089ac54bcd7da