Updated openstack/openstack Project: openstack/heat-translator  7465393c9ae90912413e2e17614a758b0e4084ea
Provide informative message for validation success

TOSCA-Parser gives errors if validation fail but for successful validation,
there is no message. Add one. Doing it separately for shell and osc as they
are having their own logic but in near future both will have a common class
not to repeat common logic and messages.

Change-Id: Ib531f315c29619ae033ce0ab13bc24adc3b90016
Closes-Bug: #1540595