Updated openstack/openstack Project: openstack/oslo.messaging  cb8b54a6fdfb6c3f79104e43b56eac8c6214b725
Adds exhange declaration on sender's side

When you send message to exchange which is not exist current
channel is closed and you need to reconnect. This is undesired.
Also this patch separate sending fanout message and call/cast
messages and don't raise exception if exchange doesn't exist
for fanout messages.

Change-Id: Ia556d0c1b219387892007925bb437664aaaccb69