Updated openstack/openstack Project: openstack/cinder  eccd7d7c5b91ee1ff7024fd7bc9199a8f4328695
EMC VMAX - not cleaning up HW Resource WWPN initiators

On cleanup, the last volume is unmapped and the masking view and
storage group are deleted. Now the initiator group and the associated
StorageHardwareIDs are also deleted as long as that initiator group
is no longer contained in another masking view.

Change-Id: Ib159f719a9c5cbac5c6ea0b7d9474a64a0335b71
Closes-Bug: #1504192