Updated openstack/openstack Project: openstack/nova  1b94d4a0ff49c3e4213c0b3f13c8f8dbacf50342
Fix two bugs in online_data_migrations()

This fixes two issues with the print statement in online_data_migrations(),
one which prevents any useful information from being displayed, and one
that actually causes us to fail in an error handling path.

The first is that print does not take a second argument of a dict for the
string formatting (like log does, which is what was probably intended here).
Thus, we dump out the raw formatted string and the dict to stdout.

The second is a formatted string in the error path for migration methods,
which wants a dict mapping but is passed a string. When that happens,
we explode.

I found these in the process of the previous patch, so decided to fix them
after making that change, but these were present in the original code.

Change-Id: I2bd0ddd1ee828a584cb062a9a6a846f5d36cb750