Updated openstack/openstack Project: openstack/neutron  6d084c339b052180f5abda4ad35e3d62e27502d8
Allow address pairs to be cleared with None

When the neutron client attempts to clear the allowed
address pairs using the '--action=clear' CLI param, it
sends a value of None to the Neutron server for allowed
address pairs. However, the allowed address pairs extension
was only allowing an empty list to clear the pairs.

This patch just converts None to an empty list on allowed
address pairs to be compatible with the client.

APIImpact
Closes-Bug: #1537734
Change-Id: Iba618b3e07bd3bdc202a9577954d7b97c2a5bf09