Updated openstack/openstack Project: openstack/nova  13d2cb8ff9236abe2f49fad9541325fd314748d2
Add build_requests database table and model

The build_requests table is a placeholder in the api database that
stores enough information about a boot request to hydrate an Instance if
necessary.  It will be populated early in the boot process so that instance
creation can be deferred until after scheduling.

Partially-implements: bp cells-scheduling-interaction
Change-Id: I8755e57c0ea9afd1c83751dea75cca640e304ab0