Updated openstack/openstack Project: openstack/nova  6a8ffe2bec70786e17bd1e664e68d25db8869a7c
Use assertRaises() to check specific exception

In test_run_terminate_deallocate_net_failure_sets_error_state(),

        ......

        try:
            self.compute.terminate_instance(self.context, instance, [], [])
        except test.TestingException:
            pass

        ......

The block aims at ensuring that terminate_instance() raises TestingException.
But this is incorrect because if TestingException is not raised by
terminate_instance(), which means if the terminate_instance() finishes normally,
the block will pass too.

This patch fixes it by using assertRaises() to check the exception.

Change-Id: I06aba3fe1199e500565d1251b180ce6ca1d4aec1