Updated openstack/openstack Project: openstack/cinder  18229c4c68c29409fb32ac24280d0cca235e4f35
Add basic workflow of attach/detach to devref

This is something that confuses folks and isn't centrally
located.  So... add it to the devref docs and hope that
developers read it.

There's tons of detail we can add in the future of course,
but this at least clears up some of the common misconceptions
that exist currently.

Change-Id: Ic5fb8da06d18fac3369a91a525accc75c1d28100