Updated openstack/openstack Project: openstack/python-neutronclient  2bf12749bc509ea8c9bf9619ac9ae814fb23b960
Ensure to use exception per status code for all cases

Previously, only when an exception has a content with
{'NeutronError': {'type': xxxx, 'message': xxxx}},
exception per status code is raised from neutronclient library.
There are cases where this kind of message is not contained
in exception messages, for example, some extension is loaded.

Library users expect an exception is raised based on response
status code and it should not depend on an exception message.
This commit applies a fallback logic to map generic per-status
exception to all exception types from the neutron server.

Closes-Bug: #1513879
Change-Id: Ib3d0a8359aed444b12217b3404d40443d61fc2c0