Updated openstack/openstack Project: openstack-infra/irc-meetings  a371d68d494981838fe0e7955e3944c5734a6fb9
Change Artifacts meeting chair & meeting ID

As Alex has moved on to another project for now, Mike is planning to
lead the meetings instead of him. Nikhil is still planning to stick
around to the project.

We are now branding Glance Artifacts as GLance Artifacts REpository or
GLARE. Hence, the follow up rename.

Change-Id: I1ff32cf80a7049aff16a8b149aacc90b4c7dda17