Updated openstack/openstack Project: openstack/cinder  d933ca7a048837f932e46fdd337c1bf06cdbe822
Return all target_wwpns for FC storwize

Storwize FC terminate_connection was using 'lsfabric' to lookup
target_wwpns, active on the fabric, to be used to terminate
the connection. However during a delete of a NPIV VM, the target_wwpns
is no longer active, it is not logged onto the fabric. The target list
will return empty. Fix is to return all the available wwpns
from 'storage_nodes'.

Change-Id: I39897f5583c878cc6243683d4bd60f800ea588af
Closes-Bug: #1546281